refactor: 성인 승객 수 조절 기능을 커스텀 훅으로 분리
