+++ src/App.tsx
@@ -1,61 +1,14 @@
-import { ChangeEvent, useState } from 'react';
 import './App.css';
-import { MESSAGE, PASSENGER } from './constants';
+import { PASSENGER } from './constants';
+import usePassengerSelect from './hooks/usePassengerSelect';
 
 function App() {
-  const [adultCount, setAdultCount] = useState<number | ''>(0);
-
-  const handleClickDecreaseAdultCount = () => {
-    if (adultCount === PASSENGER.ADULT.MIN_COUNT) {
-      alert(MESSAGE.ERROR.LESS_THAN_MIN_ADULT_COUNT);
-
-      return;
-    }
-
-    setAdultCount((prev) => {
-      if (prev === '') return PASSENGER.ADULT.MIN_COUNT;
-      return prev - 1;
-    });
-  };
-
-  const handleChangeAdultCount = ({
-    target,
-  }: ChangeEvent<HTMLInputElement>) => {
-    if (target.value === '') {
-      setAdultCount('');
-
-      return;
-    }
-
-    const value = Number(target.value);
-
-    if (value < PASSENGER.ADULT.MIN_COUNT) {
-      alert(MESSAGE.ERROR.LESS_THAN_MIN_ADULT_COUNT);
-
-      return;
-    }
-
-    if (value > PASSENGER.ADULT.MAX_COUNT) {
-      alert(MESSAGE.ERROR.EXCEED_MAX_ADULT_COUNT);
-
-      return;
-    }
-
-    setAdultCount(value);
-  };
-
-  const handleClickIncreaseAdultCount = () => {
-    if (adultCount === PASSENGER.ADULT.MAX_COUNT) {
-      alert(MESSAGE.ERROR.EXCEED_MAX_ADULT_COUNT);
-
-      return;
-    }
-
-    setAdultCount((prev) => {
-      if (prev === '') return PASSENGER.ADULT.MIN_COUNT + 1;
-      return prev + 1;
-    });
-  };
+  const {
+    adultCount,
+    handleChangeAdultCount,
+    handleClickDecreaseAdultCount,
+    handleClickIncreaseAdultCount,
+  } = usePassengerSelect();
 
   return (
     <div className='App'>
